pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-mcp-server"
version = "1.0.0"
description = "A multi-purpose MCP server with document reading, database, and project management capabilities"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.9.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "PyGithub>=2.1.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
universal-mcp-server = "universal_mcp_server.main:main"

[tool.setuptools.packages.find]
where = ["src"]